examples/requirements.txt
pillow
transformers
tf-keras
pandas
wget